requirements/base.txt
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
ecdsa==0.18.0
fastapi==0.75.2
h11==0.14.0
idna==3.4
ldap3==2.9.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
redis==4.4.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.17.1
typing_extensions==4.4.0
uvicorn==0.17.6